Completely rewrote calibration code
Curve fitting got significantly faster because now only a small window around every peak is considered. Also a new Strips class was added that provides convenient methods to perform operations on all the strips together. Overall the code should be less confusing since it was written with the exact method in mind and not explored by trial and error.